feature: add restrictions for cascading removal for admin

+++ src/main/java/com/khokhlov/tendermonitoring/controller/AdminController.java
@@ -1,13 +1,16 @@
 package com.khokhlov.tendermonitoring.controller;
 
+import com.khokhlov.tendermonitoring.model.entity.User;
 import com.khokhlov.tendermonitoring.repository.TrackedKeywordRepository;
 import com.khokhlov.tendermonitoring.repository.TrackedTenderRepository;
+import com.khokhlov.tendermonitoring.repository.UserRepository;
 import com.khokhlov.tendermonitoring.service.TrackedKeywordService;
 import com.khokhlov.tendermonitoring.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 @RequestMapping("/admin")
@@ -18,6 +21,7 @@ public class AdminController {
     private final TrackedKeywordService trackedKeywordService;
     private final TrackedKeywordRepository keywordRepository;
     private final TrackedTenderRepository trackedTenderRepository;
+    private final UserRepository userRepository;
 
     @GetMapping
     public String dashboard(Model model) {
@@ -34,7 +38,12 @@ public class AdminController {
     }
 
     @PostMapping("/users/{id}/delete")
-    public String deleteUser(@PathVariable Long id) {
+    public String deleteUser(@PathVariable Long id, RedirectAttributes redirectAttributes) {
+        User user = userRepository.getUserById(id).orElseThrow();
+        if (user.getUsername().equalsIgnoreCase("admin")) {
+            redirectAttributes.addFlashAttribute("error", "Нельзя удалить администратора.");
+            return "redirect:/admin/users";
+        }
         userService.deleteById(id);
         return "redirect:/admin/users";
     }
